Physics Facade
Drastically altered both PhysicsEngine2D, and PhysicsEngine3D so that
they are no longer self contained singletons.
PhysicsEngine acts as the singleton containing an instance of either a
2D or 3D PhysicsEngine, and using static calls to methods decides which
engine to call to, based on which engine was chosen when the
PhysicsEngine was first made.
Both 2D and 3D engines have been made internal, no one outside this
class library should actually know that either exists (because facade)