PhysicsEngine Implementation
Added PhysicsExceptions, and several exception types
Added some very basic, not even remotely implemented correctly yet
physics to the physics bodies (more as a means of making sure the whole
engine update cycle actually works)
Added RemoveBody to PhysicsBoundingChunk2D, since I apparently forgot
it, which was silly.
Added more behaviour to PhysicsEngine2D.RemoveBody() so that it removes
static bodies from the chunks they are in, ensuring there are no null
exceptions
Added chunks list to PhysicsBody2D, which gives the engine a way of
tracking what chunks have added the body to them, if the body is static,
and is a framework that will help optimise broad phase logic later for
dynamic bodies, so they can easily check against the chunks they are
already in.