UI Implementation 1.
UIBehaviour - abstract class that all different UIBehaviours will
inherit from.
UIButton - A UIVisual that will function as a button. No Implementation
yet.
UIContainer - Constructor can build a UIContainer based on the values
provided in an xml file
UIImage - A UIVisual that will draw an image. Constructor imports the
values needed through an xml file. the Draw function will use the values
from xml and will draw an image based on them.
UIManager - Will be used to manage all UIObjects. No implementation yet.
UIObject - base class for any kind of ui objects. Contains a dictionary
of UIObjects and a list of components and behaviours. UIObjects can be
anchored anywhere on the screen using the SetAlignment function and the
origin can also be set using the SetOrigin function. The ReaderTo
functions are not finished but they will import the values needed
through xml.
UIScreen - UIScreen will contain all UIObjects specific to a scene. No
implementation yet.
UISlider - A UIVisual that will function as an image slider. No
implementation yet.
UIText - A UIVisual that will draw text based on the values specifed in
an xml file. Constructor is almost finished but no other implementations
yet.
UIVisual - A UIObject that acts as the base class of all ui visuals.