UI Implementation 2.
RenderManager - Has a delegate system for each different type of
rendering we expect to have. Has a RegisterCallback method and an
UnReregisterCallBack method for Models and Sprites. The Draw Method
calls all the Draw methods registered inside each callback list.

SpriteRenderer - Contains a dictionary of Texture2Ds. A Draw Method that
uses Spritebatch to draw a Sprite based on all the parameters a UIVisual
contains. SpriteRenderers create Texture2Ds from Sprites inside the
Sprite folder using the MakeSpriteRenderer function and a SpriteRenderer
object is made out of it.

UIButton - A clone of UIImage with mouse detection for OnClick and
OnHover functionality a button would have.

Mininal Improvements in all other UI classes. Still needs lots of
revising though.