UIObject reimplementation 1
UIObject - contains a list of all child UIObjects, a name, and a
rectangle that is the bounds of the object. Contains methods to add and
remove UIObjects contained within the UIObject.

UIBehaviour - has a UIObject which is the object the behaviour is acting
on and has a duration which is the duration the behaviour will go on
for.

UIScreen - has a list of all root objects contained on a single screen.
Has methods to both add and remove objects contained within the root
object.