New UIImplementation 2
InputHandler - has a list of all InputProfiles.

InputProfile - has an enum for each type of input we expect and a player
name as well as previous and current states for each different type of
input. The update method checks the current state of each input type.

UIAlignment - has an Alignment enum for anchoring UIObjects wherever on
the screen. The GetAlignment method returns the position of the aligned
UIObject.

UIManager - Singleton class that will handle adding and removing
UIBehaviours on UIObjects. The update method calls the update method on
each uiBehaviour and removes any behaviours in the deadBehaviour list.

UIObject - now has a flag system that will be used to assign different
behaviours to a specific player profile.