Physics - Expanded Supported Collisions 2D
Collision2D
Better optimised some of the logic
Now BodyA will always be the static body if only one is
Changes to BodyA staticness allowed me to cut down the number of
condition checks.
Now Circle Circle Static Dynamic collisions are supported
Now Circle AABB Dynamic collisions have proper velocity calculation
Now Circle AABB Static Dynamic collisions are in the works

AABB
LengthAtAngle() calculates the length of a line from the origin of the
AABB to it's edge along a line of the given angle.

Circle
Changed Overlap with AABB to use the LengthAtAngle() but it broke the
collision checks, even though it shouldn't.

Ray2D
Added but realised it was pre-emptive, it will stay but I won't be doing
anything with the class for a while.