A few more render things
Font - Font uses Monogames SpriteFont class to use fonts more
extensiably. The GetFont method takes the name of a font as well as the
size you would like to draw it at. It will round to the nearest power of
2 so all fonts will be 1 of 6 different sizes. The MakeFont mewthod uses
the GetFont method to retrieve the values it needs to make a new
instance of Font.

RenderManager - Added callbacks for Text

TextRenderer - almost no implementation yet. The constructor registers
the Text_Callback and the MakeTextRenderer function makes a new
TextRenderer

UIText - almost no implementation yet.