Scenes, Assets, and Resource Management
Scene
Now has a SceneAssetsPackage

SceneAssetsPackage
Stores the string name of each asset used by this scene (have yet to
implement scene loading, and therefore have yet to implement this
package actually having any data)
Has methods that enable comparison between SceneAssetsPackages

Resources
Added LoadUIWidget() INCOMPLETE
Changed functionality of Load*() to accept the scene requesting asset
load, this can be null, this ensures that the SceneAssetsPackage of any
Scene contains all assets used by the Scene.

SceneManager
Implemented the barebones of LoadScene() this needs to be made
multi-threaded at some point in time, if it can be, or at least the
render needs to be on its own thread.