UI Prepartory Push
Resources - added 'Check' methods to find if objects exist already in
the resources loaded assets

RenderManager - added a draw callback for RenderTarget2Ds

RenderTargetRenderer

Draw() - uses the delegates draw callback to draw a rendertarget2D from
the spritebatch contined inside RenderManager.

MakeRenderTargetRenderer - checks to see if there is an instance of the
RenderTargetRenderer already with the name provided. If not it checks to
see if there is a RenderTarget at least and if there isnt it makes a new
RenderTargetRenderer with a new RenderTarget

UIText - added a few new variables and a few things to the constructor
to prepare it for the next step