RenderManager Overhaul
RenderTargetBatch - batches all objects rendering to a rendertarget
together to increase extensibility by ensuring that all similar
rendertargets are grouped together.

RenderManager - changes to the RegisterDrawCallback methods to work with
the new RenderTargetBatch system.

ModelRenderer, SpriteRenderer, TextRenderer - all have been slightly
modified to work with the new RenderTargetBatch system

WARNING: The GraphicsDevice in my version of the project seems to be
null at this point and the project will not currently build.