Custom Sprite Class and small changes
Sprite - Custom sprite class with variables for all the parameters
needed to render a Texture2D in a spritebatch

Resources - handler_Texture2D has been chaged to handler_Sprite and
accepts a template of Sprite instead of Texture2D

GraphicsHelper - now has the SpriteBatch so every class will now be
using this SpriteBatch

RenderManager - Now has a priority queue of RenderTarget callbacks. Also
has a new "screen" renderTarget that will always be rendered no matter
what.

RenderTargetBatch - now has a RenderTarget2D that is set in the draw
method.

RenderTargetRenderer - now has a RenderTarget2D that is the object to be
rendered as well as another RenderTarget2D that is the target to be
rendered to. If a target to be rendered to is null then it is passed to
be rendered to "screen"

SpriteRenderer & TextRenderer - now have a RenderTargetBatch that will
act as the batch to draw multiple sprites and text strings at once.