Prefabbing and Scene Loading - BROKEN
Scene loading now actually functions, and begins the process of
navigating through all the prefabs and assets and loading them all.

Most of the prefabbing is utterly broken, and I will switch over to
using .Net frameworks serialisation solutions instead of making my own.

For now this is a broken build.