Scene Loading and Prefabbing - Very Broken
This is all just awful.

Well Scenes aren't
Changed how they function so that the activeScene is always the scene
being switched to and not the scene being switched from
Added support for all the scene loading stuff

Problem is all the prefabbing is beyond ugly and broken.
I will be switching it all over to System.Runtime.Serialization and
using the .Net made serialisation solutions since I am breaking myself
trying to do it manually with all the runtime vs compile time typing,
and classes that don't know how to dump themselves out or come back.