Deserialisation Not Up
Work is going places.

Resources.Deserialise() is currently not working, it can't deserialise
classes from strings, but it is getting there. The next step is to get
it to make classes from their type and a value, which I will be doing by
adding a method to all the classes in question that enables them to do
it themselves, from the string in question. ({X:0 Y:1 Z:0} = new
Vector3(0,1,0))