Deserialisation Closer
Added more functionality.
It is gross. Mainly a get it up and running for Thursday kind of thing
at this point.
In the near future I will come back around and move code around until I
am comfortable with it.

As of right now the issue I see with it is that Deserialise in Resources
assumes object, and uses reflection to make said object, but then I have
had to add a whole bunch of 'special' cases in which it doesn't use
reflection and instead given the full name of the object in question
uses some other means of deserialising, and currently these are hard
coded right there, in the Deserialise method. I don't like this, it
isn't extensible, and frankly it is gross, so I need to find a way of
moving this code out of Deserialise and finding a way of teaching
objects to do this themselves, in a way that doesn't require the end
user do anything extra themselves for their own classes, so long as they
inherit from ours.

In other news, it nearly does the thing, currently for whatever reason
the object serialises fine, but comes back with only a camera, a tad
frustrating but what can I do, I will fix this later.