Deserialisation -Broken
Serialisation works, deserialisation not so much.

Solved only having the last added component on game objects, changed the
structure of adding components to use the "AddComponent" method on the
object itself, via reflection. This will work theoretically with any
object that has components (or children) so long as they all adhere to
this nomenclature, it is AddComponent(Some object base class).

It breaks still. Lots of stuff isn't getting serialised into the
.prefabs and it needs to be resolved, as well I have yet to come up with
a solution for transform heirarchies being ruined (Probably need to make
them dump out their transform parent as a ref to go find).