Title Scene Complete
Touched a whole bunch more stuff, fixed more oopsies, and we have a
lovely title/loading scene up and running, to rock peoples socks or
something.

Fundamental changes to RenderTargetBatches and RenderTargetRenderers,
such that both support RenderTargetSettings (a new class, used to be
inside RenderTargetRenderer), which greatly extends the power of both.

Fixed some pipeline issues with RenderTargetBatches and
RenderTargetRenderers (they weren't playing nice with me)

SpriteRenderers now actually listen to the alpha of colour (via
Monogames ridiculous 'premultiplied' assumptions)

A tonne of changes to the scene itself, until I was able to
aesthetically accept it for who it was.